requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.7
django-filter==2.3.0
djangorestframework==3.12.2
djangorestframework-filters==0.11.1
pkg-resources==0.0.0
pycodestyle==2.6.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
